+++ PluginsAndFeatures/azure-toolkit-for-rider/CHANGELOG.md
@@ -9,6 +9,7 @@
 - WebApp and Function property views ([#767](https://github.com/JetBrains/azure-tools-for-intellij/issues/767))
 - Azure Environment and Azure CLI path settings
 - Azure Identity settings ([#787](https://github.com/JetBrains/azure-tools-for-intellij/issues/787))
+- Option to choose Storage Account during the Function publishing ([#764](https://github.com/JetBrains/azure-tools-for-intellij/issues/764))
 
 ## [4.0.0-preview.3] - 2024-03-22